DataVis: refactored actions
